requirements.txt
catboost==1.1
joblib==1.2.0
matplotlib==3.6.1
numpy==1.23.3
optuna==3.0.3
pandas==1.5.0
pyarrow==9.0.0
PyYAML==6.0
scikit-learn==1.1.2
seaborn==0.12.0
sklearn==0.0
tqdm==4.64.1

uvicorn~=0.18.3
fastapi~=0.85.0
pydantic~=1.10.2

streamlit~=1.13.0

requests~=2.28.1
